\name{drive.file}
\alias{drive.file}
\title{Google Drive API /files "get" method}
\usage{
  drive.file(title, download.type = "txt")
}
\arguments{
  \item{download.type}{Which-ever format is appropriate for
  your document type}

  \item{title}{The exact title of the document you wish to
  download}
}
\description{
  This function searches for and retrieves files from the
  Google Drive API
}

