\name{drive.list}
\alias{drive.list}
\title{Google Drive API /files "list" method}
\usage{
  drive.list(search_term = NULL, show.titles = TRUE)
}
\arguments{
  \item{search_term}{A full title or search term}

  \item{show.titles}{If TRUE, displays a list of titles
  found by the search term}
}
\description{
  This function searches for a file or presents a list of
  files in your Google Drive
}

